requirements.txt
aiohttp
pykafka
pymongo
pyyaml
lxml
